{{cookiecutter.directory_name}}/requirements/common.txt
arrow==0.12.1
binaryornot==0.4.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
future==0.17.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.1.0
openapi-codec==1.3.2
poyo==0.4.2
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
twilio==6.20.0
uritemplate==3.0.0
urllib3==1.24.1
whichcraft==0.5.2